Extract elasticache credentials provider into its own module
This ensures it can be imported in settings
